move chat functions to firestore.ts
