+++ firebase/firestore/firestore.ts
@@ -12,6 +12,9 @@ import {
   where,
   getDocs,
   Timestamp,
+  onSnapshot,
+  serverTimestamp,
+  orderBy,
 } from "firebase/firestore";
 
 // Import type definitions used throughout the functions.
@@ -26,6 +29,9 @@ import {
   TenantCode,
   SituationReport,
 } from "../../types/types";
+
+import { auth } from "../../firebase/firebase";
+import { IMessage } from "react-native-gifted-chat";
 
 // Set the log level to 'silent' to disable logging
 // setLogLevel("silent");
@@ -626,3 +632,78 @@ export async function getSituationReportLayout(residenceId: string) {
   const situationReportLayout = residence?.situationReportLayout;
   return situationReportLayout ? situationReportLayout : [];
 }
+
+export async function sendMessage(chatId: string, content: string): Promise<void> {
+  if (!auth.currentUser) {
+    throw new Error('User must be logged in');
+  }
+
+  const user = await getUser(auth.currentUser.uid)
+
+  const chatRef = doc(db, 'chats', chatId);
+  const chat = await getDoc(chatRef);
+
+  if (!chat.exists()) {
+    throw new Error('Chat not found');
+  }
+
+  await addDoc(collection(db, 'chats', chatId, 'messages'), {
+    content: content,
+    sentBy: user?.uid,
+    sentOn: Date.now()
+  });
+}
+
+// Subscribe to messages in a chat
+export function subscribeToMessages(
+requestId: string,
+onMessagesUpdate: (messages: IMessage[]) => void
+): () => void {
+const chatRef = doc(db, 'chats', requestId);
+const messagesRef = collection(chatRef, 'messages');
+const q = query(messagesRef, orderBy('sentOn', 'desc'));
+
+// Subscribe to the query
+const unsub = onSnapshot(q, (querySnapshot) => {
+  const messages = querySnapshot.docs.map((doc) => {
+    const data = doc.data();
+    return {
+      _id: doc.id,
+      text: data["content"],
+      createdAt: data["sentOn"],
+      user: {
+        _id: data["sentBy"],
+      },
+    } as IMessage;
+  });
+
+  // Update the messages state in the View
+  onMessagesUpdate(messages);
+});
+
+// Return the unsubscribe function
+return unsub;
+}
+
+export async function createChatIfNotPresent(requestId: string): Promise<void> {
+  if (!auth.currentUser) {
+    throw new Error('User must be logged in');
+  }
+
+  const chatRef = collection(db, 'chats');
+  const chatDocRef = doc(chatRef, requestId); // Create a reference with requestId as the document ID
+
+  // Check if a chat with the given requestId already exists
+  const existingChat = await getDoc(chatDocRef);
+  if (existingChat.exists()) {
+    console.log('Chat already exists');
+    return;
+  }
+
+  // If no matching chat exists, create a new one with requestId as the document ID
+  console.log('Creating new chat');
+  await setDoc(chatDocRef, {
+    createdAt: serverTimestamp(),
+  });
+}
+
